+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - New #31: Add `sendMediaGroup` method and `InputMediaAnimation`, `InputMediaAudio`, `InputMediaDocument`,
   `InputMediaPhoto`, `InputMediaVideo` types.
 - New #32: Add `setMessageReaction` method and `ReactionType::toRequestArray()` method.
+- New #33: Add `banChatMember` and `unbanChatMember` methods.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.